+++ screens/game/playfield/painter.go
@@ -7,12 +7,12 @@ import (
 )
 
 type Painter struct {
-	Width  int16
-	Height int16
+	Width  int
+	Height int
 	Canvas *ebiten.Image
 }
 
-func newPainter(width int16, height int16) *Painter {
+func newPainter(width int, height int) *Painter {
 	return &Painter{Width: width, Height: height, Canvas: ebiten.NewImage(int(width), int(height))}
 }
 
@@ -26,14 +26,14 @@ func (this *Painter) drawGrid() {
 
 	var guideColor config.UiColor = config.GUIDE
 
-	for i := int16(0); i < rows+1; i++ {
+	for i := 0; i < rows+1; i++ {
 		lineY := i * config.SQUARE_WIDTH
 		vector.StrokeLine(
 			this.Canvas, 0, float32(lineY), float32(this.Width), float32(lineY), 1, guideColor, true,
 		)
 	}
 
-	for i := int16(0); i < columns+1; i++ {
+	for i := 0; i < columns+1; i++ {
 		lineX := i * config.SQUARE_WIDTH
 		vector.StrokeLine(
 			this.Canvas, float32(lineX), 0, float32(lineX), float32(this.Height), 1, guideColor, true,
